Created new timeVaryingMappedInletOutlet boundary condition derived from the inletOutlet mixed BC.  Code is largely based on the built-in timeVaryingMappedFixedValue BC.  The new BC has only been successfully tested with the 'fixesValues' set to true, which prevents the outflow fluxes from gbeing adjusted on these boundaries; in the adjustable flux case, the boundary flux error is extremely large and the velocity field drastically deviates from the precursor within 10-20 simulated seconds.
Additional logic for the continuity correction is supplied in a custom
adjustPhiWind routine (also part of the userfiniteVolume library).

Only one solver has been updated for now to use the new adjustPhiWind
routine instead of adjustPhi.  Currently, the solver doesn't perfectly
eliminate boundary flux error, but the fluxes do not appear to diverge.